server/requirements.txt
pyrtlsdr==0.3.0
numpy==1.26.0
websockets==12.0
